fix(lib/arch): fix MISRA C:2012 Rule 10.3 in EXTRACT usage
Also fixes Implicit conversion of "(1UL << nr_pre_bits - 5UL) - 1UL"
from essential type "unsigned 64-bit int" to different or narrower
essential type "unsigned 32-bit int".

+++ lib/arch/include/esr.h
@@ -13,7 +13,7 @@
 
 static inline unsigned int esr_srt(unsigned long esr)
 {
-	return EXTRACT(ESR_EL2_ABORT_SRT, esr);
+	return (unsigned int)EXTRACT(ESR_EL2_ABORT_SRT, esr);
 }
 
 static inline bool esr_is_write(unsigned long esr)
@@ -33,7 +33,7 @@ static inline bool esr_sixty_four(unsigned long esr)
 
 static inline unsigned int esr_sas(unsigned long esr)
 {
-	return EXTRACT(ESR_EL2_ABORT_SAS, esr);
+	return (unsigned int)EXTRACT(ESR_EL2_ABORT_SAS, esr);
 }
 
 static inline unsigned int access_len(unsigned long esr)
